requirements.txt
Django==1.10.2
argparse==1.2.1
django-geoposition==0.3.0
